+++ yash-env/src/system.rs
@@ -139,7 +139,7 @@ use self::select::SelectSystem;
 use self::select::SignalStatus;
 pub use self::shared::SharedSystem;
 pub use self::signal::{
-    CaughtSignals, Disposition, SendSignal, Sigaction, Sigmask, SigmaskOp, Signals,
+    CaughtSignals, Disposition, GetSigaction, SendSignal, Sigaction, Sigmask, SigmaskOp, Signals,
 };
 pub use self::sysconf::{ShellPath, Sysconf};
 pub use self::terminal::{Isatty, TcGetPgrp, TcSetPgrp};